consolidate the few remaining api calls
